Alteração na precisão dos pontos flutuantes de simples para dupla
Utilização de double no lugar de float para as variáveis que guardam
os coeficientes e as soluções.

Uso de ponteiro de função para o arredondamento da solução ótima para
a verificação da possibilidade de solução na função realizaPrimeiraFase.

+++ main.cpp
@@ -26,9 +26,9 @@ int main()
     std::cout << "Digite o número de restrições do problema:" << std::endl;
     std::cin >> tamanhoLinhaA;
 
-    std::vector <std::vector<float>> a(tamanhoLinhaA, std::vector<float>(tamanhoColunaA, 0));
-    std::vector<float> b(tamanhoLinhaA, 0);
-    std::vector<float> c(tamanhoColunaA, 0);
+    std::vector <std::vector<double>> a(tamanhoLinhaA, std::vector<double>(tamanhoColunaA, 0));
+    std::vector<double> b(tamanhoLinhaA, 0);
+    std::vector<double> c(tamanhoColunaA, 0);
 
     for (int i = 0 ; i < tamanhoLinhaA ; i++)
     {
